+++ .circleci/validate-all-bootstraps.sh
@@ -61,6 +61,35 @@ if ! docker login -u "${docker_username}" -p "${docker_password_DO_NOT_LOG}"; th
     exit 1
 fi
 
+# Building and running testsuites take a very long time, so we do some optimizations:
+# 1) skip building/running testsuites if only docs changes
+if git --no-pager diff --exit-code origin/develop...HEAD -- . ':!*.md' > /dev/null; then
+    echo "Skipping bootstrap validation as the only changes are in Markdown files"
+    exit 0
+fi
+
+# 2) if there are changes in the code shared across all langs, we always need to build all testsuites
+supported_langs_filepath="${root_dirpath}/${SUPPORTED_LANGS_FILENAME}"
+not_lang_dirs_filters=""
+for lang in $(cat "${supported_langs_filepath}"); do
+    not_lang_dirs_filters="${not_lang_dirs_filters} :!${lang}"
+done
+if git --no-pager diff --exit-code origin/develop...HEAD -- . ':!*.md' ${not_lang_dirs_filters} > /dev/null; then
+    has_shared_code_changes="false"
+else
+    has_shared_code_changes="true"
+fi
+
+# 3) if no shared code changes, then we only need to validate the bootstraps for the testsuites that had changes
+lang_dirs_needing_building=()
+for lang in $(cat "${supported_langs_filepath}"); do
+    if ! "${has_shared_code_changes}" && git --no-pager diff --exit-code origin/develop...HEAD -- "${lang}" > /dev/null; then
+        echo "Skipping adding ${lang} directory to list of testsuites to validate as there are no shared code changes and the directory doesn't have any changes"
+        continue
+    fi
+    lang_dirs_needing_building+=("${lang}")
+done
+
 # Git needs to be initialized, since the bootstrap will create a new Git repo and commit to it
 if ! { git config --list | grep "${GIT_USER_EMAIL_PROPERTY}"; } || ! { git config --list | grep "${GIT_USER_NAME_PROPERTY}"; }; then
     if ! git config --global "${GIT_USER_EMAIL_PROPERTY}" "bootstrap-tester@test.com"; then
@@ -74,8 +103,8 @@ if ! { git config --list | grep "${GIT_USER_EMAIL_PROPERTY}"; } || ! { git confi
 fi
 
 bootstrap_script_filepath="${root_dirpath}/${BOOTSTRAP_SCRIPTS_DIRNAME}/${BOOTSTRAP_SCRIPT_FILENAME}"
-echo "Bootstrapping and running new testsuites for all languages..."
-for lang in $(cat "${root_dirpath}/${SUPPORTED_LANGS_FILENAME}"); do
+echo "Bootstrapping and running new testsuites for languages in need of validation..."
+for lang in "${lang_dirs_needing_building[@]}"; do
     echo "Bootstrapping and running ${lang} testsuite..."
     output_dirpath="$(mktemp -d)"
     testsuite_image="bootstrap-test-${lang}-image"
@@ -93,4 +122,4 @@ for lang in $(cat "${root_dirpath}/${SUPPORTED_LANGS_FILENAME}"); do
     fi
     echo "Successfully bootstrapped and ran new ${lang} testsuite"
 done
-echo "Successfully bootstrapped and ran new testsuites for all languages!"
+echo "Successfully bootstrapped and ran new testsuites for all languages in need of validation!"